requirements.txt
psycopg2-binary
langchain
langchain-aws
python-dotenv
langchain-chroma
langchain-community
